add PhotosGridWithPagination component
