+++ src/components/Pagination.scss
@@ -13,6 +13,6 @@
   }
 
   &__current-page {
-    font-size: 1.2rem;
+    font-size: 1.4rem;
   }
 }
