+++ src/components/PhotosGridWithPagination.scss
@@ -1,0 +1,5 @@
+.photos-with-pagination {
+  display: flex;
+  flex-direction: column;
+  gap: 2.5rem;
+}
